+++ src/batch-loader/batch-loader.ts
@@ -1,54 +1,9 @@
 import { BatchAggregator } from '../batch-aggregator/batch-aggregator'
 import { BatchLoaderFn } from '../batch-aggregator/interfaces'
 import { Deduplicator } from '../deduplicator/deduplicator'
-import { Key } from '../interfaces'
-import { CacheAdapter, ICache } from './cache-adapter'
-
-export interface IBatchLoaderOptions<K, R> {
-  /**
-   * @description Function to extract the key from a query
-   * @default query => `${query}`
-   */
-  getKey?: (query: K) => Key
-
-  cache?: ICache<R>
-
-  /**
-   * @description Task execution timeout in milliseconds
-   * @default 60_000
-   */
-  timeoutMs?: number
-
-  /**
-   * @description Allows timers to avoid blocking the event loop
-   * @default false
-   */
-  unrefTimeouts?: boolean
-
-  /**
-   * @description Maximum number of parallel tasks (default: unlimited)
-   * @default Infinity
-   */
-  concurrencyLimit?: number
-
-  /**
-   * @description Maximum number of requests per batch
-   * @default 1000
-   */
-  maxBatchSize?: number
-
-  /**
-   * @description Maximum time in milliseconds to form a batch
-   * @default 50
-   */
-  batchTimeMs?: number
-
-  /**
-   * @description Maximum waiting time in milliseconds for tasks in the queue (only if concurrencyLimit > 0)
-   * @default 60_000
-   */
-  maxWaitingTimeMs?: number
-}
+import { Key } from '../utils/interfaces'
+import { CacheAdapter } from './cache-adapter'
+import { IBatchLoaderOptions } from './interfaces'
 
 const prepareOptions = <K, R>(options: IBatchLoaderOptions<K, R>) => ({
   getKey: (query: K) => `${query}`,
